Find the largest number
Find the largest number that can be formed with the given digits

Given an array of integers arr[] representing digits of a number. The task is to write a program to generate the largest number possible using these digits.
Note: The digits in the array are in between 0 and 9. That is, 0<arr[i]<9.
Examples:

Input : arr[] = {4, 7, 9, 2, 3}
Output : Largest number: 97432

Input : arr[] = {8, 6, 0, 4, 6, 4, 2, 7}
Output : Largest number: 87664420